% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivation.R
\name{estimate}
\alias{estimate}
\title{Extract the point estimate from a derivation}
\usage{
estimate(x)
}
\arguments{
\item{x}{An \code{epibyhand_derivation}.}
}
\value{
Numeric scalar.
}
\description{
Extract the point estimate from a derivation
}
