% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi2x2.R
\name{print.epi2x2}
\alias{print.epi2x2}
\title{Print a 2 x 2 table with margins}
\usage{
\method{print}{epi2x2}(x, indent = "", ...)
}
\arguments{
\item{x}{An \code{epi2x2}.}

\item{indent}{Character prefix for each line.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a 2 x 2 table with margins
}
