% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{print.epi_strata}
\alias{print.epi_strata}
\title{Print stratified tables}
\usage{
\method{print}{epi_strata}(x, indent = "", ...)
}
\arguments{
\item{x}{An \code{epi_strata}.}

\item{indent}{Character prefix for each line.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print stratified tables
}
